fixed lots of tests
