+++ impl/src/test/kotlin/nl/nuts/consent/bridge/pipelines/CordaStateChangeListenerControllerTest.kt.tmp
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package nl.nuts.consent.bridge.listener
+package nl.nuts.consent.bridge.pipelines
 
 import com.nhaarman.mockito_kotlin.any
 import com.nhaarman.mockito_kotlin.eq
