+++ impl/src/test/kotlin/nl/nuts/consent/bridge/pipelines/CordaStateChangeListenerIntegrationTest.kt.tmp
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package nl.nuts.consent.bridge.listener
+package nl.nuts.consent.bridge.pipelines
 
 import com.natpryce.hamkrest.assertion.assertThat
 import com.natpryce.hamkrest.contains
@@ -28,7 +28,6 @@ import net.corda.client.rpc.CordaRPCClient
 import net.corda.client.rpc.CordaRPCClientConfiguration
 import net.corda.client.rpc.CordaRPCConnection
 import net.corda.core.contracts.StateAndRef
-import net.corda.core.internal.concurrent.map
 import net.corda.core.messaging.startFlow
 import net.corda.node.services.Permissions
 import net.corda.testing.core.ALICE_NAME
@@ -40,7 +39,6 @@ import nl.nuts.consent.bridge.Serialization
 import nl.nuts.consent.bridge.nats.Event
 import nl.nuts.consent.bridge.nats.EventName
 import nl.nuts.consent.bridge.nats.EventStateStore
-import nl.nuts.consent.bridge.nats.NutsEventListenerTest
 import nl.nuts.consent.bridge.nats.NutsEventPublisher
 import nl.nuts.consent.bridge.rpc.CordaRPClientWrapper
 import nl.nuts.consent.bridge.rpc.test.DummyFlow
@@ -48,7 +46,6 @@ import nl.nuts.consent.bridge.rpc.test.DummyFlow.ConsumeFlow
 import nl.nuts.consent.bridge.rpc.test.DummyFlow.ProduceFlow
 import nl.nuts.consent.bridge.rpc.test.DummyState
 import np.com.madanpokharel.embed.nats.EmbeddedNatsConfig
-import np.com.madanpokharel.embed.nats.EmbeddedNatsServer
 import np.com.madanpokharel.embed.nats.NatsServerConfig
 import np.com.madanpokharel.embed.nats.NatsStreamingVersion
 import np.com.madanpokharel.embed.nats.ServerType
@@ -56,10 +53,8 @@ import org.junit.*
 import java.io.File
 import java.util.*
 import java.util.concurrent.CountDownLatch
-import java.util.concurrent.atomic.AtomicInteger
 import java.util.concurrent.atomic.AtomicReference
 import kotlin.test.BeforeTest
-import kotlin.test.assertEquals
 import kotlin.test.assertNotNull
 
 /**
